refactor: move SteveConfigurationReader to utils package and update imports
