+++ steve-ui-jsp/src/main/java/de/rwth/idsg/steve/web/BatchInsertConverter.java
