+++ steve/src/main/java/de/rwth/idsg/steve/Application.java
@@ -19,6 +19,7 @@
 package de.rwth.idsg.steve;
 
 import de.rwth.idsg.steve.utils.LogFileRetriever;
+import de.rwth.idsg.steve.utils.SteveConfigurationReader;
 import lombok.extern.slf4j.Slf4j;
 
 import java.nio.file.Path;
