+++ steve/src/main/java/de/rwth/idsg/steve/utils/SteveConfigurationReader.java
@@ -16,9 +16,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package de.rwth.idsg.steve;
+package de.rwth.idsg.steve.utils;
 
-import de.rwth.idsg.steve.utils.PropertiesFileLoader;
+import de.rwth.idsg.steve.ApplicationProfile;
+import de.rwth.idsg.steve.SteveConfiguration;
 import lombok.experimental.UtilityClass;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
