+++ steve/src/test/java/de/rwth/idsg/steve/ApplicationTest.java
@@ -18,6 +18,7 @@
  */
 package de.rwth.idsg.steve;
 
+import de.rwth.idsg.steve.utils.SteveConfigurationReader;
 import de.rwth.idsg.steve.utils.__DatabasePreparer__;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterAll;
