+++ steve/src/test/java/de/rwth/idsg/steve/StressTestJsonOCPP16.java
@@ -20,6 +20,7 @@ package de.rwth.idsg.steve;
 
 import de.rwth.idsg.steve.ocpp.OcppVersion;
 import de.rwth.idsg.steve.utils.OcppJsonChargePoint;
+import de.rwth.idsg.steve.utils.SteveConfigurationReader;
 import de.rwth.idsg.steve.utils.StressTester;
 import lombok.RequiredArgsConstructor;
 import ocpp.cs._2015._10.AuthorizationStatus;
