+++ steve/src/test/java/de/rwth/idsg/steve/StressTestSoapOCPP16.java
@@ -18,6 +18,7 @@
  */
 package de.rwth.idsg.steve;
 
+import de.rwth.idsg.steve.utils.SteveConfigurationReader;
 import de.rwth.idsg.steve.utils.StressTester;
 import lombok.RequiredArgsConstructor;
 import ocpp.cs._2015._10.AuthorizationStatus;
