+++ steve/src/test/java/de/rwth/idsg/steve/issues/Issue72LowLevelSoap.java
@@ -19,7 +19,7 @@
 package de.rwth.idsg.steve.issues;
 
 import com.google.common.net.MediaType;
-import de.rwth.idsg.steve.SteveConfigurationReader;
+import de.rwth.idsg.steve.utils.SteveConfigurationReader;
 import de.rwth.idsg.steve.StressTest;
 import de.rwth.idsg.steve.utils.Helpers;
 import de.rwth.idsg.steve.utils.StressTester;
